requirements.txt
python-twitter
requests
facebook-sdk